+++ src/detectors/compose/index.ts
@@ -1,0 +1,2 @@
+export { composeLargeComposableDetector } from "./largeComposable.js";
+export { composeStateHoistingDetector } from "./stateHoisting.js";